+++ test/logster/test_logger.rb
@@ -61,6 +61,20 @@ class TestLogger < Minitest::Test
     end
   end
 
+  class NewLogger < Logster::Logger; end
+
+  def test_inherited_logger_backtrace_with_chain
+    @other_store = TestStore.new
+    @logger = NewLogger.new(@store)
+    @logger.chain(Logster::Logger.new(@other_store))
+
+    @logger.add(0, "test", "prog", backtrace: "backtrace", env: { a: "x" })
+
+    [@store, @other_store].each do |store|
+      assert_equal "backtrace", store.calls[0][3][:backtrace]
+    end
+  end
+
   def test_progname_parameter
     @logger.add(0, "test")
     progname = @store.calls[0][1]